EIPS eligible for Inclusion (#2378)
* EIPS eligible for Inclusion

Meta EIP for tracking Eligibility for Inclusion as defined in the EIP-centric Forking Model.

* Spelling and formatting updates

* Formating Github Markdown

* Added number and responded to Feedback

* Update and rename EIP-2378.md to eip-2378.md

* EFI: Updated Discussion-to URL to the /EIP

Given that this will be a registry EIP and the informational part will be moved out eventually. It made sense to me for the appropriate discussion about what EIPs are included should happen on the /ethereum/eip gitter channel.